Separated custom location manager
